requirements.txt
whisper-ai
sounddevice
numpy
